Contrib from Jiri Novak <novak@physics.muni.cz> and Karolína Pěčková <437271@mail.muni.cz> report on bugs regarding the parameters uncertainties 2020-01-28 11:11 Dear Emmanuel, we are writing regarding a series of bugs in iFit concerning uncertainties output and calculations.
The bugs concern:
1) the output of the the parameters uncertainties
2) the calculation of uncertainties from Hessian matrix
3) the calculation of uncertainties from parameters and criterion
trajectory (history)

Please, find attached a detailed report (the pdf) on the reported bugs
and their demonstration,
example matlab scripts and functions and results of tests
(example_uncertainty_output_bug.m and uncertainty_estimation.zip)
and proposed fixes of the iFit functions (fixes.zip).

Regarding the calculation of the parameters uncertainties from the
parameters and criterion trajectory,
we are not sure about a correct fix. We wanted to ask you whether the
formula implemented in fmin_private_wrapper.m
(which is actually different from that in the iFit manual) is based on
some particular textbook or paper on statistical
data analysis? If it is so, please, could you kindly provide a
reference? Also, please, did you or somebody from the users
do some (thorough) testing on uncertainty estimators?

Thank you in advance for your answer.

Answer: thanks for contribution. Jiri and Karolina did not manage to push a PR.
I thus commit on their behalf.